% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{SF_palette}
\alias{SF_palette}
\title{A Space Movie palette generator ....}
\usage{
SF_palette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of desired palette. See `index`(\url{http://}) for a full list of package colours.}

\item{n}{Number of colours required.}

\item{type}{"Continuous" for interpolated palettes, or "discrete".}
}
\value{
A vector of colours.
}
\description{
These are a handful of color palettes from the Space Movie franchise.
}
\examples{
SW_palette("ANH")
SW_palette("ROTS", 3)

If you need more colors than normally found in a palette, you can use a
continuous palette to interpolate between existing colors.

pal <- SF_palette(name = "AOTC", n = 21, type = "continuous")
image(volcano, col = pal)
}
\keyword{colours}

