% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{SF_palettes}
\alias{SF_palettes}
\title{Complete list of palettes}
\format{An object of class \code{list} of length 12.}
\usage{
SF_palettes
}
\description{
Use \code{\link{SF_palette}} to construct palettes of desired length.
}
\keyword{datasets}

