% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfPalettes.R
\docType{package}
\name{sfPalettes}
\alias{sfPalettes}
\alias{sfPalettes-package}
\title{sfPalettes - Colour palettes based on sci-fi movies}
\description{
Colour palettes based on sci-fi movies
}
\details{
Colour palettes based on sci-fi movies
}

