% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SF_palette.R
\name{SF_palette}
\alias{SF_palette}
\title{SF_palette}
\usage{
SF_palette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of desired palette.}

\item{n}{Number of colours required.}

\item{type}{"Continuous" for interpolated palettes, or "discrete".}
}
\value{
A vector of colours.
}
\description{
Use \code{\link{SF_palette}} to construct palettes of desired length.
}
\examples{
SW_palette("ANH")
SW_palette("JP", 3)

Palettes can be interpolated, increasing the number of available colours. 

pal <- SF_palette(name = "HER", n = 21, type = "continuous")
image(volcano, col = pal)

}
\seealso{
Karthik Ram's `wesanderson`(\url{https://github.com/karthik/wesanderson}) color palettes.
}
\keyword{colours}

