% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfpalettes.R
\docType{package}
\name{sfPalettes}
\alias{sfPalettes}
\alias{sfPalettes-package}
\title{sfPalettes}
\description{
list of palettes based on science fiction movies
}
\details{
list of palettes based on science fiction movies
}

